requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

altair==5.5.0
click==8.2.1
numpy==1.26.4
pandas==2.3.1
pyarrow==21.0.0
pyfredapi==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.0
six==1.17.0
